Add simple feature-gating
