+++ tests/error/nightly/derives_for_generic_structs_with_backtrace.rs
