+++ tests/error/nightly/derives_for_generic_structs_with_source_backtrace_chaining.rs
