+++ tests/error/nightly/derives_for_structs_with_backtrace.rs
