+++ tests/error_tests.rs
@@ -1,4 +1,4 @@
-#![feature(backtrace)]
+#![cfg_attr(feature = "nightly", feature(backtrace))]
 
 #[macro_use]
 extern crate derive_more;
